Move scripts into packages, create unit test workflow
* Change file structure to use python module/package conventions
* Create Makefile with cleanup and test tasks
* Create example unit test
