+++ .gitignore
@@ -1,4 +1,3 @@
 **.retry
-etl-scripts/config.py
-etl-scripts/__pycache__
-etl-scripts/__pycache__/*
+__pycache__
+.cache
