+++ docs/Quasar-Architecture-v0.1.png
