+++ docs/Quasar-Architecture-v0.2.png
