+++ quasar/etl/CampaignActivityFullBackfill.py
@@ -1,6 +1,6 @@
 import logging
 
-from QuasarCampaignActivityParsing import RogueEtl
+from .QuasarCampaignActivityParsing import RogueEtl
 
 if __name__ == "__main__":
     log_format = "%(asctime)s - %(levelname)s: %(message)s"
