+++ quasar/etl/DSMySQL.py
@@ -1,7 +1,7 @@
 import MySQLdb
 
 import config
-from DSHelper import QuasarException
+from .DSHelper import QuasarException
 
 # MySQL Connector Library for Blade Platform.
