+++ quasar/etl/DSNorthstarScraper.py
@@ -3,7 +3,7 @@ import config
 import oauthlib
 from oauthlib.oauth2 import BackendApplicationClient
 from requests_oauthlib import OAuth2Session
-from QuasarWebScraper import Scraper
+from .QuasarWebScraper import Scraper
 
 class NorthstarScraper(Scraper):
     """Class for extracting Northstar data via API."""
