+++ quasar/etl/QuasarCustomerIOProcessQueue.py
@@ -1,4 +1,4 @@
-from QuasarCustomerIOQueue import QuasarQueue
+from .QuasarCustomerIOQueue import QuasarQueue
 import logging
