+++ quasar/etl/thor_mailchimp_subscriber_to_user_table.py
