+++ quasar/misc/QuasarQueue-Furnes-Improvements.py
